requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

anyio==3.7.1
async-timeout==4.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pylast==5.2.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
urllib3==2.0.4
